+++ clocket@tirtha/files/clocket@tirtha/CHANGELOG.md
@@ -84,4 +84,5 @@ Changelog
   - hide decorations (set to true by default)
   - show clock setting
   - show date setting
-  - clock custom time string
+  - custom clock time string
+  - custom date time string